Fix terminal persistence and thread turn handling
Apply the verified review findings across dispatcher, thread, worker,
testing, and webhook code paths.

- preserve structured tool results instead of double-encoding JSON strings
- split turn execution helpers into focused sibling modules
- make terminal job result persistence atomic with status updates
- tighten hydration, compaction, auth-cancellation, and rollback behaviour
- refresh testing docs, null-db helpers, and webhook lifecycle guidance
