+++ src/agent/dispatcher/delegate/mod.rs
@@ -23,9 +23,14 @@ use crate::context::JobContext;
 use crate::error::Error;
 use crate::llm::{Reasoning, ReasoningContext};
 
-// Re-export items used by other modules in the crate
-// These are used by tests and other modules, but not within this module
-#[allow(unused_imports)]
+// Re-export items used by other modules in the crate.
+#[cfg_attr(
+    not(test),
+    expect(
+        unused_imports,
+        reason = "re-exported for external modules/tests; used outside this module"
+    )
+)]
 pub(crate) use llm_hooks::{compact_messages_for_retry, strip_internal_tool_call_text};
 pub(crate) use tool_exec::{
     ToolCallSpec, check_auth_required, execute_chat_tool_standalone, parse_auth_result,
