+++ tests/support/webhook_helpers.rs
@@ -15,7 +15,6 @@ use serde_json::json;
 use ironclaw::channels::{WebhookServer, WebhookServerConfig};
 
 /// A started webhook server with a `/health` route and a pre-built client.
-#[allow(dead_code, reason = "consumed selectively across test binaries")]
 pub struct StartedWebhookServer {
     pub server: WebhookServer,
     pub addr: SocketAddr,
@@ -23,13 +22,11 @@ pub struct StartedWebhookServer {
 }
 
 /// Return the standard `/health` check route used by webhook tests.
-#[allow(dead_code, reason = "consumed selectively across test binaries")]
 pub fn health_routes() -> Router {
     Router::new().route("/health", get(|| async { Json(json!({"status": "ok"})) }))
 }
 
 /// Build a reqwest client with the standard 2-second test timeout.
-#[allow(dead_code, reason = "consumed selectively across test binaries")]
 pub fn test_http_client() -> Result<reqwest::Client, reqwest::Error> {
     reqwest::Client::builder()
         .timeout(Duration::from_secs(2))
@@ -39,7 +36,6 @@ pub fn test_http_client() -> Result<reqwest::Client, reqwest::Error> {
 /// Bind an ephemeral listener, build a WebhookServer with a `/health`
 /// route, start the server, and return the started server plus a
 /// preconfigured client.
-#[allow(dead_code, reason = "consumed selectively across test binaries")]
 pub async fn start_health_server()
 -> Result<StartedWebhookServer, Box<dyn std::error::Error + Send + Sync>> {
     let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await?;
